# orchestrator/docs/swagger.yaml
basePath: /
definitions:
  handler.signInInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.ComputingResource:
    properties:
      id:
        type: string
      last_ping_at:
        type: string
      work_state:
        type: string
    type: object
  models.ExpressionFromUser:
    properties:
      expression:
        type: string
    type: object
  models.ExpressionToRead:
    properties:
      created_at:
        type: string
      expression:
        type: string
      id:
        type: string
      result:
        type: string
      solved_at:
        type: string
      work_state:
        type: string
    type: object
  models.Operation:
    properties:
      divide:
        type: string
      minus:
        type: string
      multiply:
        type: string
      plus:
        type: string
    type: object
  models.User:
    properties:
      name:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - name
    - password
    - username
    type: object
host: localhost:8080
info:
  contact:
    email: faridsadreddinov@yandex.ru
    name: Фарид
    url: https://t.me/M00nfI0wer
  description: Это простой распределенный калькулятор.
  title: Distributed Calculator
  version: "1.0"
paths:
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: Login
      parameters:
      - description: credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.signInInput'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Sign in
      tags:
      - auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: Create accout
      parameters:
      - description: User info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: integer
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Sign up
      tags:
      - auth
  /computing_resources/:
    get:
      consumes:
      - application/json
      description: Get info about computing resources
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ComputingResource'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get computing resources
      tags:
      - computing resources
  /expressions/:
    get:
      consumes:
      - application/json
      description: Get expressions
      parameters:
      - description: Page num
        in: query
        name: page
        required: true
        type: integer
      - description: Num of record per page
        in: query
        name: recordPerPage
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ExpressionToRead'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get expressions
      tags:
      - expressions
    post:
      consumes:
      - application/json
      description: Add new expression
      parameters:
      - description: Expression info
        in: body
        name: expression
        required: true
        schema:
          $ref: '#/definitions/models.ExpressionFromUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Add expression
      tags:
      - expressions
  /expressions/{id}:
    get:
      consumes:
      - application/json
      description: Get expression by id
      parameters:
      - description: Expression ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ExpressionToRead'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get expression
      tags:
      - expressions
  /operations/:
    get:
      consumes:
      - application/json
      description: Get time of operations in seconds
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Operation'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get time of operations
      tags:
      - operations
    patch:
      consumes:
      - application/json
      description: Update time of operations in seconds
      parameters:
      - description: Time of operations
        in: body
        name: operations
        required: true
        schema:
          $ref: '#/definitions/models.Operation'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Operation'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update operation time
      tags:
      - operations
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"